# Make CQL executor service usage optional [cql-tests]