--- a/janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/CQLColValGetter.java
+++ b/janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/CQLColValGetter.java
@@ -20,7 +20,7 @@ import org.janusgraph.diskstorage.EntryMetaData;
 import org.janusgraph.diskstorage.StaticBuffer;
 import org.janusgraph.diskstorage.util.StaticArrayEntry.GetColVal;
 
-class CQLColValGetter implements GetColVal<Tuple3<StaticBuffer, StaticBuffer, Row>, StaticBuffer> {
+public class CQLColValGetter implements GetColVal<Tuple3<StaticBuffer, StaticBuffer, Row>, StaticBuffer> {
 
     private final EntryMetaData[] schema;