--- a/janusgraph-hbase/src/main/java/org/janusgraph/diskstorage/hbase/HBaseStoreManager.java
+++ b/janusgraph-hbase/src/main/java/org/janusgraph/diskstorage/hbase/HBaseStoreManager.java
@@ -464,7 +464,7 @@ public class HBaseStoreManager extends DistributedStoreManager implements KeyCol
         }
 
         if (commitTime != null) {
-            sleepAfterWrite(txh, commitTime);
+            sleepAfterWrite(commitTime);
         }
     }