… docs/advanced-topics/executor-service.md
… docs/configs/janusgraph-cfg.md
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/CQLTransaction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/ColumnOperationFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/ConsumerWithBackendException.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/AbstractCQLMutateManyFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/AbstractCQLMutateManyLoggedFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/AbstractCQLMutateManyUnloggedFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/CQLExecutorServiceMutateManyLoggedFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/CQLExecutorServiceMutateManyUnloggedFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/CQLMutateManyFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/CQLSimpleMutateManyLoggedFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/mutate/CQLSimpleMutateManyUnloggedFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/slice/AbstractCQLSliceFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/slice/CQLExecutorServiceSliceFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/slice/CQLSimpleSliceFunction.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/function/slice/CQLSliceFunction.java
… janusgraph-cql/src/test/java/org/janusgraph/diskstorage/cql/CachingCQLStoreManager.java
… janusgraph-cql/src/test/java/org/janusgraph/graphdb/cql/CQLGraphTest.java