This PR adds possibility to work with CQL without ExecutorService at all. I.e. relying on `MAX_REQUESTS_PER_CONNECTION` and internal DataStax pools instead of relying on the default or custom pool.
Potentially working with disabled CQL ExecutorService should be the same or better in performance and have less overhead (i.e. less unnecessary created threads).
If we prove that working with disabled ExecutorService is better in terms of memory management and / or performance, we can disable CQL ExecutorService by default in JanusGraph and potentially deprecate / remove the usage of unnecessary CQL ExecutorService.

Performance tests conducted:
https://github.com/JanusGraph/janusgraph/pull/2741#issuecomment-894331436
https://github.com/JanusGraph/janusgraph/pull/2741#issuecomment-894635159
https://github.com/JanusGraph/janusgraph/pull/2741#issuecomment-894656248
https://github.com/JanusGraph/janusgraph/pull/2741#issuecomment-894663680

Before opening this PR I also checked that `CQLGraphTest` tests passes with disabled ExecutorService by default (i.e. tests in `JanusGraphTest`) but if we test all `CQLGraphTest` tests with enabled and disabled `ExecutorService` it will double the execution time for CQL tests. Thus, I simply extended `testConsistencyEnforcement` and `testConcurrentConsistencyEnforcement` to run their parametrized tests with both enabled and disabled `ExecutorService`. I don't know, but maybe we need to add a flow to trigger all CQL tests with disabled `ExecutorService` when we create a commit with `[cql-tests]`. 

Related discussions with @li-boxuan and @cdegroc:
https://github.com/JanusGraph/janusgraph/pull/2700#discussion_r669480986
https://github.com/JanusGraph/janusgraph/discussions/2545#discussioncomment-551453

Fixes #2707

Signed-off-by: Oleksandr Porunov <alexandr.porunov@gmail.com>

-----

Thank you for contributing to JanusGraph!

In order to streamline the review of the contribution we ask you
to ensure the following steps have been taken:

### For all changes:
- [x] Is there an issue associated with this PR? Is it referenced in the commit message?
- [x] Does your PR body contain #xyz where xyz is the issue number you are trying to resolve?
- [x] Has your PR been rebased against the latest commit within the target branch (typically `master`)?
- [x] Is your initial contribution a single, squashed commit?

### For code changes:
- [x] Have you written and/or updated unit tests to verify your changes?
- [ ] If adding new dependencies to the code, are these dependencies licensed in a way that is compatible for inclusion under [ASF 2.0](https://www.apache.org/legal/resolved.html#category-a)?
- [ ] If applicable, have you updated the LICENSE.txt file, including the main LICENSE.txt file in the root of this repository?
- [ ] If applicable, have you updated the NOTICE.txt file, including the main NOTICE.txt file found in the root of this repository?

### For documentation related changes:
- [x] Have you ensured that format looks appropriate for the output in which it is rendered?