@cdegroc 2021-08-04
Thanks for doing this @porunov 👍  I'm really curious to see the performance differences!
If the goal is to merge this as-is and not just run performance tests, my main concern would be about code duplication and maintainability.