@porunov 2021-08-04
@cdegroc Thanks for the review! I applied your suggestions and tried to reduce code duplication more.