@porunov 2021-08-04
> Thanks for doing this @porunov +1 I'm really curious to see the performance differences!

The main purpose of this PR is to be able to run the same CQL queries without using potentially unnecessary threads. I don't think there will be much performance difference between enabled and disabled ExecutorService.
The only difference I could think of is if you have high throughput ability on your server but you configured high `storage.cql.max-requests-per-connection` and low `storage.cql.executor-service.max-pool-size`. If so, yes, there should theoretically be performance difference. Other than that, performance improvement could also come from less context switches between different threads when we are not using additional ExecutorService (i.e. we simply have less unnecessary threads on a server) but I wouldn't give that much expectations either. 

That said, I didn't really conduct and performance testing.