@porunov 2021-08-06
Made very simple benchmark test which simply adds 100,000 empty vertices, 2000 vertices with a property and 1000 edges with a property.
After that it traverses all vertices and gets properties for these vertices as well as edges for the vertices and properties of the edges.
The test with enabled by default CQL executor service is located here: https://gist.github.com/porunov/ce3422a8266d9feb4c89452ff705b9ce
The test with disabled CQL executor service is located here (the same test but with added `configuration.set(EXECUTOR_SERVICE_ENABLED.toStringWithoutRoot(), false);`): https://gist.github.com/porunov/9b0c10ec0c06c1ba32695897db9b2412

The results are below:
```
VM version: JDK 11.0.12, OpenJDK 64-Bit Server VM, 11.0.12+7
```

**On master branch with enabled CQL executor service**:
```
Benchmark                             (assignTimestamp)  (atomicBatch)  (batchSize)  Mode  Cnt      Score   Error  Units
TraverseBenchmark.testTraversingTime               true           true           20  avgt       49474.682          ms/op
TraverseBenchmark.testTraversingTime               true           true            1  avgt       49918.248          ms/op
TraverseBenchmark.testTraversingTime               true          false           20  avgt       42622.825          ms/op
TraverseBenchmark.testTraversingTime               true          false            1  avgt       44450.675          ms/op
TraverseBenchmark.testTraversingTime              false           true           20  avgt       44478.228          ms/op
TraverseBenchmark.testTraversingTime              false           true            1  avgt       40274.519          ms/op
TraverseBenchmark.testTraversingTime              false          false           20  avgt       41803.870          ms/op
TraverseBenchmark.testTraversingTime              false          false            1  avgt       41076.673          ms/op
```

**On this branch with enabled by default CQL executor service**:
```
Benchmark                             (assignTimestamp)  (atomicBatch)  (batchSize)  Mode  Cnt      Score   Error  Units
TraverseBenchmark.testTraversingTime               true           true           20  avgt       40880.514          ms/op
TraverseBenchmark.testTraversingTime               true           true            1  avgt       41069.039          ms/op
TraverseBenchmark.testTraversingTime               true          false           20  avgt       41636.535          ms/op
TraverseBenchmark.testTraversingTime               true          false            1  avgt       41224.890          ms/op
TraverseBenchmark.testTraversingTime              false           true           20  avgt       40837.060          ms/op
TraverseBenchmark.testTraversingTime              false           true            1  avgt       41950.444          ms/op
TraverseBenchmark.testTraversingTime              false          false           20  avgt       40490.052          ms/op
TraverseBenchmark.testTraversingTime              false          false            1  avgt       41422.679          ms/op
```

**On this branch with disabled CQL executor service**:
```
Benchmark                             (assignTimestamp)  (atomicBatch)  (batchSize)  Mode  Cnt      Score   Error  Units
TraverseBenchmark.testTraversingTime               true           true           20  avgt       36145.111          ms/op
TraverseBenchmark.testTraversingTime               true           true            1  avgt       35734.069          ms/op
TraverseBenchmark.testTraversingTime               true          false           20  avgt       36026.026          ms/op
TraverseBenchmark.testTraversingTime               true          false            1  avgt       36430.664          ms/op
TraverseBenchmark.testTraversingTime              false           true           20  avgt       35902.646          ms/op
TraverseBenchmark.testTraversingTime              false           true            1  avgt       35687.994          ms/op
TraverseBenchmark.testTraversingTime              false          false           20  avgt       36476.968          ms/op
TraverseBenchmark.testTraversingTime              false          false            1  avgt       35778.240          ms/op
```

As seen above, the execution time for tests on `master` branch and this branch with enabled by default CQL executor service is similar in performance. The execution time for tests with disabled CQL executor service on this branch is lower for all the cases tested.
Thus, I can confirm that working with disabled executor service may be better in performance in some situations. I can't say in all situations yet because I simply didn't test all the cases but probably most of the cases will benefit from disabled CQL executor service (at least we can see about 10% increase in performance for this case on my machine). 
Notice, this test is executing benchmarks in one thread, thus it doesn't test the multi threaded access by multiple clients to the same `CQLStoreManager`. Probably, to test that case we could remove `RUN_COUNT` (to always return the same `keyspace` configuration) and add additional check to `setup` method which will not create schema & data if schema is already initialized.

P.S. I didn't expect to see 10% improvement for the test. I think it could be just a bad test. More various tests are needed to confirm the performance improvement.