@porunov 2021-08-07
Simple write test: https://gist.github.com/porunov/6a6f1b0a5dd654871714f663d4a83c51
To test disabled executor service you need to add `configuration.set(EXECUTOR_SERVICE_ENABLED.toStringWithoutRoot(), false);` to `setup` method.

Results are below.

**Master branch**:
```
Benchmark                     (assignTimestamp)  (atomicBatch)  (batchSize)  Mode  Cnt    Score      Error  Units
WriteBenchmark.testWriteTime               true           true           20  avgt    3  820.602 ± 1239.062  ms/op
WriteBenchmark.testWriteTime               true           true            1  avgt    3  742.753 ± 2447.104  ms/op
WriteBenchmark.testWriteTime               true          false           20  avgt    3  730.592 ± 2318.229  ms/op
WriteBenchmark.testWriteTime               true          false            1  avgt    3  826.049 ± 4052.735  ms/op
WriteBenchmark.testWriteTime              false           true           20  avgt    3  777.851 ± 2107.037  ms/op
WriteBenchmark.testWriteTime              false           true            1  avgt    3  776.049 ± 2353.500  ms/op
WriteBenchmark.testWriteTime              false          false           20  avgt    3  714.974 ± 2048.363  ms/op
WriteBenchmark.testWriteTime              false          false            1  avgt    3  838.614 ± 5144.424  ms/op
```

**This branch with enabled CQL executor service by default**:
```
Benchmark                     (assignTimestamp)  (atomicBatch)  (batchSize)  Mode  Cnt    Score      Error  Units
WriteBenchmark.testWriteTime               true           true           20  avgt    3  750.470 ± 1792.536  ms/op
WriteBenchmark.testWriteTime               true           true            1  avgt    3  902.898 ± 6193.790  ms/op
WriteBenchmark.testWriteTime               true          false           20  avgt    3  702.975 ± 2318.588  ms/op
WriteBenchmark.testWriteTime               true          false            1  avgt    3  840.229 ± 5032.483  ms/op
WriteBenchmark.testWriteTime              false           true           20  avgt    3  835.900 ± 4783.879  ms/op
WriteBenchmark.testWriteTime              false           true            1  avgt    3  899.230 ± 6222.718  ms/op
WriteBenchmark.testWriteTime              false          false           20  avgt    3  736.021 ± 2091.344  ms/op
WriteBenchmark.testWriteTime              false          false            1  avgt    3  819.863 ± 4776.885  ms/op
```

**This branch with disabled CQL executor service**:
```
Benchmark                     (assignTimestamp)  (atomicBatch)  (batchSize)  Mode  Cnt    Score      Error  Units
WriteBenchmark.testWriteTime               true           true           20  avgt    3  883.683 ± 6222.681  ms/op
WriteBenchmark.testWriteTime               true           true            1  avgt    3  761.552 ± 1324.556  ms/op
WriteBenchmark.testWriteTime               true          false           20  avgt    3  687.512 ± 1920.683  ms/op
WriteBenchmark.testWriteTime               true          false            1  avgt    3  733.218 ± 3038.014  ms/op
WriteBenchmark.testWriteTime              false           true           20  avgt    3  766.956 ± 2653.458  ms/op
WriteBenchmark.testWriteTime              false           true            1  avgt    3  878.886 ± 6276.718  ms/op
WriteBenchmark.testWriteTime              false          false           20  avgt    3  683.059 ± 2509.469  ms/op
WriteBenchmark.testWriteTime              false          false            1  avgt    3  682.410 ± 2408.270  ms/op
```

Conclusion about the "write" test: Write tests didn't show performance differences between `master` branch with enabled CQL executor service, this branch with enabled executor service and this branch with disabled executor service.