@porunov 2021-08-07
Multi threaded test: https://gist.github.com/porunov/0b01f62da2440f9567c0a47e82c0785d

I was comparing the result of the test to this one: https://github.com/JanusGraph/janusgraph/pull/2741#issuecomment-894331436

Both tests do the same (i.e. same data is created and same data is traversed).

Notice, I tested only the first case when `assignTimestamp = true, atomicBatch = true, batchSize = 20`.  To test with all cases, change the test parameters to include all values (with all parameters the amount of tests will be `48`).

As always, to test disable executor service you need to add `configuration.set(EXECUTOR_SERVICE_ENABLED.toStringWithoutRoot(), false);` to `setup` method.

The results are the next:

**Master branch (enabled CQL executor service)**:
```
Benchmark                                          (assignTimestamp)  (atomicBatch)  (batchSize)  (threadsCount)  Mode  Cnt       Score   Error  Units
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               1  avgt        40961.136          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               3  avgt        37205.040          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               5  avgt        44785.057          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              10  avgt        51406.648          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              20  avgt        80802.159          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              50  avgt       120320.723          ms/op
```

**This branch with enabled CQL executor service by default**:

```
Benchmark                                          (assignTimestamp)  (atomicBatch)  (batchSize)  (threadsCount)  Mode  Cnt       Score   Error  Units
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               1  avgt        41784.622          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               3  avgt        37371.193          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               5  avgt        45240.449          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              10  avgt        51302.146          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              20  avgt        77971.718          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              50  avgt       120231.977          ms/op
```

**This branch with disabled CQL executor service**:
```
Benchmark                                          (assignTimestamp)  (atomicBatch)  (batchSize)  (threadsCount)  Mode  Cnt      Score   Error  Units
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               1  avgt       36841.976          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               3  avgt       42757.243          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20               5  avgt       43562.073          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              10  avgt       45961.538          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              20  avgt       54717.432          ms/op
TraverseMultithreadedBenchmark.testTraversingTime               true           true           20              50  avgt       90342.676          ms/op
```

As seen above, again, `master` branch and this branch are the same in performance when executor service is enabled (default case).
When it comes to disabled CQL executor service, we see performance improvement.
There was a statistical anomaly in the second case when the amount of threads were `3`. The execution time had to be similar or bigger than the execution time with `1` thread but was faster. Thus, I think we can disregard that case.